+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # HEAD
 
+- Added `makePolDegreeMap`, `makePolAngMap` and `differenceAngMaps`.
 - Added `makeErroredMap` and `makeIdealMap` to generate sky maps observed by a given telescope setup.